ato_children/api/views.py
from django.contrib.auth.models import User
from rest_framework import viewsets

from ..models import Gift, Volunteer, City

from .serializers import (UserSerializer, GiftSerializer,
                          VolunteerSerializer, CitySerializer)


# ViewSets define the view behavior.
class UserViewSet(viewsets.ModelViewSet):
    queryset = User.objects.all()
    serializer_class = UserSerializer


class GiftViewSet(viewsets.ModelViewSet):
    queryset = Gift.objects.all()
    serializer_class = GiftSerializer


class VolunteeViewSet(viewsets.ModelViewSet):
    queryset = Volunteer.objects.all()
    serializer_class = VolunteerSerializer


class CityViewSet(viewsets.ModelViewSet):
    queryset = City.objects.all()
    serializer_class = CitySerializer
