requirements.txt
dj-database-url==0.3.0

Django==1.9

gunicorn==19.3.0
whitenoise==2.0.4
psycopg2==2.5.4

ujson==1.33
djangorestframework==3.3.1
Markdown==2.6.5
django-filter==0.11.0
django-crispy-forms==1.5.2
